Separated and decoupled the ViewModel logic for each screen, so they each have their own ViewModel responsible for their own states and logic. RecipeListView has RecipeListViewModel, and RecipeDetailView has RecipeDetailViewModel. Shared state logic was moved into the Repository, as it is a Singleton and injected into both ViewModels, hence providing each one with access. Selected recipes are now stored in the repository, as is the list of recipes to be viewed. This was done due to the structure of the data being too complex to be passed with a navigation event, and the data lacking any unique ID that could otherwise be used to reference which Recipe was being selected (titles are not reliable unique IDs, for example).
Additionally, navigation events from the ListView to the DetailView are now triggered by pressing on the card, which triggers the ViewModel to update a UserEvent StateFlow with the intent, which flows back to the View and triggers a navigation event. This way the ViewModel is responsible, rather than the event management happening all in the View.

Finally, minor changes to Composable signatures occurred to remove unnecessary parameters such as navController and imports. RecipeDetailView's padding was also adjusted to be consistent with RecipeListView.